requirements.txt
streamlit
pandas
matplotlib
plotly
numpy
bcrypt

streamlit
mysql-connector-python
